+++ packages/@expo/prebuild-config/src/plugins/unversioned/expo-splash-screen/getAndroidSplashConfig.ts
@@ -1,13 +1,13 @@
 import { ExpoConfig } from '@expo/config-types';
 
 export type SplashScreenConfig = {
-  xxxhdpi?: string | null;
-  xxhdpi?: string | null;
-  xhdpi?: string | null;
-  hdpi?: string | null;
-  mdpi?: string | null;
-  image?: string | null;
-  backgroundColor: string | null;
+  xxxhdpi?: string;
+  xxhdpi?: string;
+  xhdpi?: string;
+  hdpi?: string;
+  mdpi?: string;
+  image?: string;
+  backgroundColor?: string;
   resizeMode: 'contain' | 'cover' | 'native';
   dark?: {
     backgroundColor?: string;
@@ -36,12 +36,12 @@ export function getAndroidSplashConfig(
   if (props) {
     const splash = props;
     return {
-      xxxhdpi: splash.xxxhdpi ?? splash.image ?? null,
-      xxhdpi: splash.xxhdpi ?? splash.image ?? null,
-      xhdpi: splash.xhdpi ?? splash.image ?? null,
-      hdpi: splash.hdpi ?? splash.image ?? null,
-      mdpi: splash.mdpi ?? splash.image ?? null,
-      backgroundColor: splash.backgroundColor ?? null,
+      xxxhdpi: splash.xxxhdpi ?? splash.image,
+      xxhdpi: splash.xxhdpi ?? splash.image,
+      xhdpi: splash.xhdpi ?? splash.image,
+      hdpi: splash.hdpi ?? splash.image,
+      mdpi: splash.mdpi ?? splash.image,
+      backgroundColor: splash.backgroundColor,
       resizeMode: splash.resizeMode ?? defaultResizeMode,
       imageWidth: splash.imageWidth,
     };
@@ -50,12 +50,13 @@ export function getAndroidSplashConfig(
   if (config.android?.splash) {
     const splash = config.android?.splash;
     return {
-      xxxhdpi: splash.xxxhdpi ?? splash.image ?? null,
-      xxhdpi: splash.xxhdpi ?? splash.image ?? null,
-      xhdpi: splash.xhdpi ?? splash.image ?? null,
-      hdpi: splash.hdpi ?? splash.image ?? null,
-      mdpi: splash.mdpi ?? splash.image ?? null,
-      backgroundColor: splash.backgroundColor ?? null,
+      xxxhdpi: splash.xxxhdpi ?? splash.image,
+      xxhdpi: splash.xxhdpi ?? splash.image,
+      xhdpi: splash.xhdpi ?? splash.image,
+      hdpi: splash.hdpi ?? splash.image,
+      mdpi: splash.mdpi ?? splash.image,
+      backgroundColor: splash.backgroundColor,
+      image: splash.image,
       resizeMode: splash.resizeMode ?? defaultResizeMode,
       imageWidth: 200,
     };
@@ -64,12 +65,12 @@ export function getAndroidSplashConfig(
   if (config.splash) {
     const splash = config.splash;
     return {
-      xxxhdpi: splash.image ?? null,
-      xxhdpi: splash.image ?? null,
-      xhdpi: splash.image ?? null,
-      hdpi: splash.image ?? null,
-      mdpi: splash.image ?? null,
-      backgroundColor: splash.backgroundColor ?? null,
+      xxxhdpi: splash.image,
+      xxhdpi: splash.image,
+      xhdpi: splash.image,
+      hdpi: splash.image,
+      mdpi: splash.image,
+      backgroundColor: splash.backgroundColor,
       resizeMode: splash.resizeMode ?? defaultResizeMode,
       imageWidth: 200,
     };
@@ -80,18 +81,18 @@ export function getAndroidSplashConfig(
 
 export function getAndroidDarkSplashConfig(
   config: Pick<ExpoConfig, 'splash' | 'android'>,
-  props: AndroidSplashConfig | null
+  props?: AndroidSplashConfig | null
 ): SplashScreenConfig | null {
   if (props?.dark) {
     const splash = props.dark;
     const lightTheme = getAndroidSplashConfig(config, props);
     return {
-      xxxhdpi: splash.xxxhdpi ?? splash.image ?? null,
-      xxhdpi: splash.xxhdpi ?? splash.image ?? null,
-      xhdpi: splash.xhdpi ?? splash.image ?? null,
-      hdpi: splash.hdpi ?? splash.image ?? null,
-      mdpi: splash.mdpi ?? splash.image ?? null,
-      backgroundColor: splash.backgroundColor ?? null,
+      xxxhdpi: splash.xxxhdpi ?? splash.image,
+      xxhdpi: splash.xxhdpi ?? splash.image,
+      xhdpi: splash.xhdpi ?? splash.image,
+      hdpi: splash.hdpi ?? splash.image,
+      mdpi: splash.mdpi ?? splash.image,
+      backgroundColor: splash.backgroundColor,
       resizeMode: lightTheme?.resizeMode ?? defaultResizeMode,
     };
   }
@@ -102,12 +103,12 @@ export function getAndroidDarkSplashConfig(
     const splash = config.android?.splash?.dark;
     const lightTheme = getAndroidSplashConfig(config, props);
     return {
-      xxxhdpi: splash.xxxhdpi ?? splash.image ?? null,
-      xxhdpi: splash.xxhdpi ?? splash.image ?? null,
-      xhdpi: splash.xhdpi ?? splash.image ?? null,
-      hdpi: splash.hdpi ?? splash.image ?? null,
-      mdpi: splash.mdpi ?? splash.image ?? null,
-      backgroundColor: splash.backgroundColor ?? null,
+      xxxhdpi: splash.xxxhdpi ?? splash.image,
+      xxhdpi: splash.xxhdpi ?? splash.image,
+      xhdpi: splash.xhdpi ?? splash.image,
+      hdpi: splash.hdpi ?? splash.image,
+      mdpi: splash.mdpi ?? splash.image,
+      backgroundColor: splash.backgroundColor,
       // Can't support dark resizeMode because the resize mode is hardcoded into the MainActivity.java
       resizeMode: lightTheme?.resizeMode ?? defaultResizeMode,
     };