+++ packages/@expo/prebuild-config/src/plugins/unversioned/expo-splash-screen/withIosSplashScreenStoryboardImage.ts
@@ -27,6 +27,7 @@ export function applySplashScreenStoryboard(obj: IBSplashScreenDocument, splash:
       contentMode,
       imageName,
       backgroundColor: splash.backgroundColor,
+      enableFullScreenImage: splash.enableFullScreenImage_legacy ?? false,
       imageWidth: splash.imageWidth,
     });
   }